+++ BuildMasterAutomation/Functions/Invoke-BMRestMethod.ps1
@@ -159,9 +159,11 @@ function Invoke-BMRestMethod
     {
         if( $Method -eq [Microsoft.PowerShell.Commands.WebRequestMethod]::Get -or $PSCmdlet.ShouldProcess($Uri,$Method) )
         {
+            Write-Debug "$($webRequestParam.ToString())"
             Invoke-RestMethod -Method $Method -Uri $uri @webRequestParam -Headers $headers |
                 ForEach-Object { $_ } |
                 Where-Object { $_ }
+
         }
     }
     catch